OT98-62 fix comments
