+++ src/main/java/com/alkemy/ong/controller/ContactController.java
@@ -1,7 +1,6 @@
 package com.alkemy.ong.controller;
 
 import com.alkemy.ong.common.converter.ConvertUtils;
-import com.alkemy.ong.exception.SendEmailException;
 import com.alkemy.ong.model.entity.Contact;
 import com.alkemy.ong.model.request.CreateContactRequest;
 import com.alkemy.ong.model.response.DetailsContactResponse;
@@ -35,7 +34,7 @@ public class ContactController {
       produces = MediaType.APPLICATION_JSON_VALUE,
       consumes = MediaType.APPLICATION_JSON_VALUE)
   public ResponseEntity<DetailsContactResponse> create(
-      @RequestBody CreateContactRequest createContactRequest) throws SendEmailException {
+      @RequestBody CreateContactRequest createContactRequest) {
     Contact contact = createContactService.create(createContactRequest);
     DetailsContactResponse createContactResponse = this.convertUtils.toResponse(contact);
     return new ResponseEntity<>(createContactResponse, HttpStatus.CREATED);
