+++ src/main/java/com/alkemy/ong/service/ContactServiceImpl.java
@@ -4,6 +4,7 @@ import com.alkemy.ong.common.converter.ConvertUtils;
 import com.alkemy.ong.common.mail.EmailHelper;
 import com.alkemy.ong.exception.SendEmailException;
 import com.alkemy.ong.model.entity.Contact;
+import com.alkemy.ong.model.request.ContactTemplateEmail;
 import com.alkemy.ong.model.request.CreateContactRequest;
 import com.alkemy.ong.model.response.DetailsContactResponse;
 import com.alkemy.ong.model.response.ListContactResponse;
@@ -12,10 +13,12 @@ import com.alkemy.ong.service.abstraction.ICreateContactService;
 import com.alkemy.ong.service.abstraction.IListContactsService;
 import java.util.Date;
 import java.util.List;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
+@Slf4j
 public class ContactServiceImpl implements ICreateContactService, IListContactsService {
 
   private static final Date ACTIVE_CONTACT = null;
@@ -30,14 +33,20 @@ public class ContactServiceImpl implements ICreateContactService, IListContactsS
   EmailHelper emailHelper;
 
   @Override
-  public Contact create(CreateContactRequest createContactRequest) throws SendEmailException {
+  public Contact create(CreateContactRequest createContactRequest) {
     Contact contact = new Contact();
     contact.setName(createContactRequest.getName());
     contact.setPhone(createContactRequest.getPhone());
     contact.setEmail(createContactRequest.getEmail());
     contact.setMessage(createContactRequest.getMessage());
     contact.setDeletedAt(ACTIVE_CONTACT);
-    emailHelper.send(createContactRequest);
+    try {
+      ContactTemplateEmail contactTemplateEmail = new ContactTemplateEmail();
+      contactTemplateEmail.setEmail(createContactRequest.getEmail());
+      emailHelper.send(contactTemplateEmail);
+    } catch (SendEmailException s) {
+      log.info(s.getMessage());
+    }
     return contactRepository.save(contact);
   }
